Move MailchimpBot to Mailchimp::Bot service
