+++ app/services/moderator/manage_activity_and_roles.rb
@@ -35,14 +35,14 @@ module Moderator
       @user.remove_role :trusted
       @user.remove_role :tag_moderator
       @user.update(email_tag_mod_newsletter: false)
-      MailchimpBot.new(user).manage_tag_moderator_list
+      Mailchimp::Bot.new(user).manage_tag_moderator_list
       @user.update(email_community_mod_newsletter: false)
-      MailchimpBot.new(user).manage_community_moderator_list
+      Mailchimp::Bot.new(user).manage_community_moderator_list
     end
 
     def remove_tag_moderator_role
       @user.remove_role :tag_moderator
-      MailchimpBot.new(user).manage_tag_moderator_list
+      Mailchimp::Bot.new(user).manage_tag_moderator_list
     end
 
     def create_note(reason, content)
