+++ app/services/tag_moderators/add.rb
@@ -54,7 +54,7 @@ module TagModerators
       Rails.cache.delete("user-#{user.id}/tag_moderators_list")
       return unless tag_mod_newsletter_enabled?
 
-      MailchimpBot.new(user).manage_tag_moderator_list
+      Mailchimp::Bot.new(user).manage_tag_moderator_list
     end
 
     def tag_mod_newsletter_enabled?
